Added One Sentence per line rule (#66)
New configurable and fixable rule to fix ensure only one sentence
per line. It is configurable, what defines the ending of a
sentence as well as ignored words.

Additionally to the new rule, i adapted the logic to create the
yaml file, as it did not work properly with arrays.

Signed-off-by: Simon Schrottner <simon.schrottner@dynatrace.com>